Adjusted the way group image update works. Handled some cases with conflicting file ids and updated the upload logic. #773

+++ app/src/main/java/com/clover/studio/exampleapp/data/models/UploadFile.kt
@@ -11,7 +11,7 @@ data class UploadFile(
     val offset: Long,
 
     // total: Total number of chunks
-    val total: Long,
+    val total: Int,
 
     // size: File size in bytes
     val size: Long,
